OverridingUnionTypeResolverTrait
