+++ layers/CMSSchema/packages/customposts-wp/src/Overrides/TypeResolvers/UnionType/CustomPostUnionTypeResolver.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace PoPCMSSchema\CustomPostsWP\Overrides\TypeResolvers\UnionType;
 
 use PoPCMSSchema\CustomPosts\TypeResolvers\UnionType\CustomPostUnionTypeResolver as UpstreamCustomPostUnionTypeResolver;
-use PoPCMSSchema\SchemaCommons\Overrides\TypeResolvers\OverridingTypeResolverTrait;
+use PoPCMSSchema\SchemaCommons\Overrides\TypeResolvers\OverridingUnionTypeResolverTrait;
 
 class CustomPostUnionTypeResolver extends UpstreamCustomPostUnionTypeResolver
 {
-    use OverridingTypeResolverTrait;
+    use OverridingUnionTypeResolverTrait;
 }
